BT_28 Removed collision from individual parts of Tank_BP to prevent bug, Added quit button to the menu

+++ BattleTank/Source/BattleTank/Public/Tank.h
@@ -56,5 +56,5 @@ private:
 
 	UTankBarrel* Barrel = nullptr;
 	
-	double LastFireTime = 0;
+	float LastFireTime = 0.f;
 };
